+++ tests/training/test_utils.py
@@ -11,7 +11,8 @@ import torch
 from pykeen.losses import MarginRankingLoss
 from pykeen.models import Model, TransE
 from pykeen.training.lcwa import LCWATrainingLoop
-from pykeen.training.utils import apply_label_smoothing, lazy_compile_random_batches
+from pykeen.training.utils import lazy_compile_random_batches
+from pykeen.utils import apply_label_smoothing
 from pykeen.triples import TriplesFactory
